requirements.txt
# Web3 and Blockchain
web3>=6.0.0
eth-account>=0.10.0

# Data handling
pandas>=2.0.0

# Configuration and environment
python-dotenv>=1.0.0
pyyaml>=6.0

# Utilities
colorama>=0.4.6

# GUI dependencies
pystray>=0.19.0
Pillow>=10.0.0
APScheduler>=3.10.0
pywin32>=306  # Windows DPAPI for encrypted key storage

# Build tools (optional, for creating .exe)
# pyinstaller>=6.0.0
